Feature: Added new transformer and updated existing (#32)
In the latest PR, the Account REPO was fixed. Features added include a Progression commit, working controllers, endpoint messages, and account activation. Scopes were incorporated for better access control. The system now uses the transformers library with the class-validator/transformer for improved data handling. Testing was expanded with hour frequency checks and a single-item test. Additionally, a new string-to-number transformer was introduced and existing ones updated.
---------

Co-authored-by: Tomáš Ječmínek <jecminek.tomas@outlook.com>
Co-authored-by: Tomáš Ječmínek <jecminek@navisys.cz>